requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
python-multipart==0.0.9
pillow==10.4.0
numpy
google-cloud-aiplatform==1.73.0
opencv-python==4.10.0.84
pytesseract==0.3.10
importlib-metadata>=6.7.0; python_version < "3.10"
torch
torchvision
easyocr==1.7.2
google-generativeai
